fix bpf compilation
